requirements.txt
Flask==1.1.1
gunicorn==20.0.4
python-dotenv==0.12.0
pytube3==9.6.4
requests==2.23.0
psycopg2-binary==2.8.5
Flask-SQLAlchemy==2.4.1
